# server/docs/swagger.yaml
definitions:
...
  file.FileInfo:
    properties:
      content:
        type: string
      extension:
        type: string
      favoriteID:
        type: integer
      gid:
        type: string
      group:
        type: string
      isDir:
        type: boolean
      isHidden:
        type: boolean
      isSymlink:
        type: boolean
      itemTotal:
        type: integer
      items:
        items:
          $ref: '#/definitions/file.FileInfo'
        type: array
      linkPath:
        type: string
      mimeType:
        type: string
      modTime:
        type: string
      mode:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      type:
        type: string
      uid:
        type: string
      updateTime:
        type: string
      user:
        type: string
    type: object
...
  request.FileCreate:
    properties:
      content:
        type: string
      isDir:
        type: boolean
      isLink:
        type: boolean
      isSymlink:
        type: boolean
      linkPath:
        type: string
      mode:
        type: integer
      path:
        type: string
      sub:
        type: boolean
    required:
    - mode
    - path
    type: object
  request.FileOption:
    properties:
      containSub:
        type: boolean
      dir:
        type: boolean
      expand:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      path:
        type: string
      search:
        type: string
      showHidden:
        type: boolean
      sortBy:
        type: string
      sortOrder:
        type: string
    type: object
  request.FilePathCheck:
    properties:
      path:
        type: string
    required:
    - path
    type: object
...
  response.FileInfo:
    properties:
      content:
        type: string
      extension:
        type: string
      favoriteID:
        type: integer
      gid:
        type: string
      group:
        type: string
      isDir:
        type: boolean
      isHidden:
        type: boolean
      isSymlink:
        type: boolean
      itemTotal:
        type: integer
      items:
        items:
          $ref: '#/definitions/file.FileInfo'
        type: array
      linkPath:
        type: string
      mimeType:
        type: string
      modTime:
        type: string
      mode:
        type: string
      name:
        type: string
      path:
        type: string
      size:
        type: integer
      type:
        type: string
      uid:
        type: string
      updateTime:
        type: string
      user:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.5.9
paths:
...
  /files:
    post:
      consumes:
      - application/json
      description: 创建文件/文件夹
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileCreate'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Create file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Create dir or file [path]
        formatZH: 创建文件/文件夹 [path]
        paramKeys: []
  /files/check:
    post:
      consumes:
      - application/json
      description: 检测文件是否存在
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FilePathCheck'
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Check file exist
      tags:
      - File
  /files/search:
    post:
      consumes:
      - application/json
      description: 获取文件列表
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.FileOption'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.FileInfo'
      security:
      - ApiKeyAuth: []
      summary: List files
      tags:
      - File
  /files/upload:
    post:
      description: 上传文件
      parameters:
      - description: request
        in: formData
        name: file
        required: true
        type: file
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Upload file
      tags:
      - File
      x-panel-log:
        BeforeFunctions: []
        bodyKeys:
        - path
        formatEN: Upload file [path]
        formatZH: 上传文件 [path]
        paramKeys: []
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"